# Acorn-based parsing