--- a/Jint.Tests/Runtime/Debugger/TestHelpers.cs
+++ b/Jint.Tests/Runtime/Debugger/TestHelpers.cs
@@ -9,7 +9,7 @@ namespace Jint.Tests.Runtime.Debugger
             return node switch
             {
                 Directive directive => requiredValue == null || directive.Value == requiredValue,
-                ExpressionStatement expr => requiredValue == null || (expr.Expression is Literal literal && literal.StringValue == requiredValue),
+                NonSpecialExpressionStatement expr => requiredValue == null || (expr.Expression is StringLiteral literal && literal.Value == requiredValue),
                 _ => false
             };
         }