--- a/Jint/Messages.cs
+++ b/Jint/Messages.cs
@@ -1,9 +1,0 @@
-namespace Jint;
-
-/// <summary>
-/// Helper to map Esprima error messages.
-/// </summary>
-internal static class Messages
-{
-    public const string InvalidLHSInAssignment = "Invalid left-hand side in assignment";
-}