… Jint.Tests.Test262/Test262Harness.settings.json
… Jint.Tests/Runtime/ParserOptionsPropagationTests.cs
… Jint/AstExtensions.cs
… Jint/Engine.Ast.cs
… Jint/Engine.cs
… Jint/Native/Function/ClassDefinition.cs
… Jint/Native/Function/EvalFunction.cs
… Jint/Native/Generator/GeneratorInstance.cs
… Jint/Native/Global/GlobalObject.cs
… Jint/Native/RegExp/RegExpConstructor.cs
… Jint/Native/ShadowRealm/ShadowRealm.cs
… Jint/ParsingOptions.cs
… Jint/Runtime/CallStack/CallStackElement.cs
… Jint/Runtime/CallStack/JintCallStack.cs
… Jint/Runtime/Completion.cs
… Jint/Runtime/Debugger/BreakLocation.cs
… Jint/Runtime/Debugger/CallFrame.cs
… Jint/Runtime/Debugger/DebugHandler.cs
… Jint/Runtime/Debugger/DebugInformation.cs
… Jint/Runtime/Environments/FunctionEnvironment.cs
… Jint/Runtime/Interpreter/EvaluationContext.cs
… Jint/Runtime/Interpreter/Expressions/DestructuringPatternAssignmentExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintAssignmentExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintBinaryExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintLiteralExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintLogicalAndExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintLogicalOrExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintMemberExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintObjectExpression.cs