@adams85 2024-03-29
BenchmarkDotNet v0.13.12, Windows 10 (10.0.19045.4170/22H2/2022Update)
AMD Ryzen 7 2700X, 1 CPU, 16 logical and 8 physical cores
.NET SDK 8.0.202
  [Host]     : .NET 8.0.3 (8.0.324.11423), X64 RyuJIT AVX2
  DefaultJob : .NET 8.0.3 (8.0.324.11423), X64 RyuJIT AVX2

**Before (Esprima)**

| Method            | FileName             | Mean             | StdDev         | Rank | Allocated      |
|------------------ |--------------------- |-----------------:|---------------:|-----:|---------------:|
| Jint              | array-stress         |     7,515.224 μs |     42.4682 μs |    1 |     7017.84 KB |
| Jint_ParsedScript | array-stress         |     8,157.611 μs |    115.4234 μs |    2 |     6996.66 KB |
|                   |                      |                  |                |      |                |
| Jint              | dromaeo-3d-cube      |    28,390.069 μs |     42.0761 μs |    4 |     6241.11 KB |
| Jint_ParsedScript | dromaeo-3d-cube      |    27,196.405 μs |     34.6054 μs |    3 |      5983.8 KB |
|                   |                      |                  |                |      |                |
| Jint              | dromaeo-core-eval    |     4,727.442 μs |     13.2962 μs |    3 |      355.79 KB |
| Jint_ParsedScript | dromaeo-core-eval    |     4,475.208 μs |     13.5812 μs |    2 |      338.75 KB |
|                   |                      |                  |                |      |                |
| Jint              | dromaeo-object-array |    53,183.217 μs |    784.0281 μs |    1 |   100410.77 KB |
| Jint_ParsedScript | dromaeo-object-array |    56,760.258 μs |    930.0847 μs |    2 |   100370.95 KB |
|                   |                      |                  |                |      |                |
| Jint              | droma(...)egexp [21] |   229,861.280 μs |  2,785.7601 μs |    2 |   163402.02 KB |
| Jint_ParsedScript | droma(...)egexp [21] |   222,161.400 μs |  2,183.8414 μs |    1 |   169953.97 KB |
|                   |                      |                  |                |      |                |
| Jint              | droma(...)tring [21] |   404,473.279 μs |  3,376.5817 μs |    1 |  1321472.77 KB |
| Jint_ParsedScript | droma(...)tring [21] |   398,929.113 μs |  5,418.5646 μs |    1 |  1321261.68 KB |
|                   |                      |                  |                |      |                |
| Jint              | droma(...)ase64 [21] |    60,223.052 μs |     63.6666 μs |    3 |      6124.7 KB |
| Jint_ParsedScript | droma(...)ase64 [21] |    59,242.062 μs |    120.8232 μs |    2 |     6039.83 KB |
|                   |                      |                  |                |      |                |
| Jint              | evaluation           |        36.788 μs |      0.6763 μs |    2 |       40.85 KB |
| Jint_ParsedScript | evaluation           |        15.226 μs |      0.1624 μs |    1 |       32.28 KB |
|                   |                      |                  |                |      |                |
| Jint              | linq-js              |     2,483.586 μs |      6.9467 μs |    3 |     1273.64 KB |
| Jint_ParsedScript | linq-js              |       141.609 μs |      1.8270 μs |    1 |      224.88 KB |
|                   |                      |                  |                |      |                |
| Jint              | minimal              |         5.802 μs |      0.2169 μs |    3 |        14.6 KB |
| Jint_ParsedScript | minimal              |         3.164 μs |      0.0050 μs |    1 |       13.22 KB |
|                   |                      |                  |                |      |                |
| Jint              | stopwatch            |   389,799.457 μs |  2,555.6312 μs |    4 |    53042.25 KB |
| Jint_ParsedScript | stopwatch            |   385,872.907 μs |  2,871.8493 μs |    4 |    53017.58 KB |Before (Esprima):

** After (Acornima) **


| Method            | FileName             | Mean             | StdDev         | Rank | Allocated      |
|------------------ |--------------------- |-----------------:|---------------:|-----:|---------------:|
| Jint              | array-stress         |     7,493.907 μs |     55.6531 μs |    1 |     7016.32 KB |
| Jint_ParsedScript | array-stress         |     8,521.961 μs |    111.3640 μs |    3 |     6994.32 KB |
|                   |                      |                  |                |      |                |
| Jint              | dromaeo-3d-cube      |    28,166.099 μs |     27.7357 μs |    4 |     6234.68 KB |
| Jint_ParsedScript | dromaeo-3d-cube      |    26,540.061 μs |     88.2389 μs |    3 |     5981.46 KB |
|                   |                      |                  |                |      |                |
| Jint              | dromaeo-core-eval    |     4,642.193 μs |     22.5442 μs |    3 |       357.3 KB |
| Jint_ParsedScript | dromaeo-core-eval    |     4,494.251 μs |      7.5824 μs |    2 |      339.72 KB |
|                   |                      |                  |                |      |                |
| Jint              | dromaeo-object-array |    52,620.869 μs |  1,219.8760 μs |    1 |      100411 KB |
| Jint_ParsedScript | dromaeo-object-array |    53,944.009 μs |  1,143.7667 μs |    2 |    100368.6 KB |
|                   |                      |                  |                |      |                |
| Jint              | droma(...)egexp [21] |   233,301.670 μs |  3,384.7220 μs |    2 |   164630.45 KB |
| Jint_ParsedScript | droma(...)egexp [21] |   225,059.003 μs |  2,321.0998 μs |    1 |   165143.63 KB |
|                   |                      |                  |                |      |                |
| Jint              | droma(...)tring [21] |   403,190.871 μs |  5,494.1274 μs |    1 |  1321527.96 KB |
| Jint_ParsedScript | droma(...)tring [21] |   397,381.593 μs |  5,916.8257 μs |    1 |  1321348.63 KB |
|                   |                      |                  |                |      |                |
| Jint              | droma(...)ase64 [21] |    59,188.533 μs |    142.2596 μs |    2 |     6119.13 KB |
| Jint_ParsedScript | droma(...)ase64 [21] |    59,294.962 μs |    126.4828 μs |    2 |     6036.86 KB |
|                   |                      |                  |                |      |                |
| Jint              | evaluation           |        32.682 μs |      0.0295 μs |    2 |       38.71 KB |
| Jint_ParsedScript | evaluation           |        14.002 μs |      0.0705 μs |    1 |       29.95 KB |
|                   |                      |                  |                |      |                |
| Jint              | linq-js              |     2,405.574 μs |      2.3482 μs |    3 |     1327.49 KB |
| Jint_ParsedScript | linq-js              |       135.846 μs |      2.2389 μs |    1 |      222.54 KB |
|                   |                      |                  |                |      |                |
| Jint              | minimal              |         4.376 μs |      0.0120 μs |    2 |       13.45 KB |
| Jint_ParsedScript | minimal              |         3.029 μs |      0.0046 μs |    1 |       12.09 KB |
|                   |                      |                  |                |      |                |
| Jint              | stopwatch            |   382,592.787 μs |  3,267.6964 μs |    4 |    53041.01 KB |
| Jint_ParsedScript | stopwatch            |   377,164.633 μs |  1,199.8201 μs |    4 |    53015.24 KB |