@ejsmith 2024-03-30
Would this give Jint better error messages with line and column information?